requirements.txt
Flask==2.3.3
crewai==0.28.8
langchain-openai==0.1.8
requests==2.31.0
python-dotenv==1.0.0